examples/requirements/django_3_1_and_elastic_7x.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile django_3_1_and_elastic_7x.in
#
asgiref==3.4.1
    # via django
attrs==21.2.0
    # via pytest
backports.entry-points-selectable==1.1.0
    # via virtualenv
certifi==2021.5.30
    # via elasticsearch
coverage[toml]==5.5
    # via
    #   -r test.in
    #   pytest-cov
distlib==0.3.2
    # via virtualenv
django==3.1.13
    # via
    #   -r django_3_1.in
    #   django-cors-headers
    #   django-debug-toolbar
    #   django-nine
    #   djangorestframework
django-cors-headers==3.6.0
    # via -r django_3_1.in
django-debug-toolbar==3.2.1
    # via -r django_3_1.in
django-debug-toolbar-force==0.1.8
    # via -r django_3_1.in
django-elasticsearch-debug-toolbar==2.0.0
    # via -r django_3_1.in
django-elasticsearch-dsl==7.2.2
    # via -r elastic_7x.in
django-nine==0.2.4
    # via
    #   django-debug-toolbar-force
    #   django-ormex
django-ormex==0.2.1
    # via -r django_3_1.in
djangorestframework==3.12.4
    # via -r django_3_1.in
elasticsearch==7.10.1
    # via
    #   -r elastic_7x.in
    #   elasticsearch-dsl
elasticsearch-dsl==7.3.0
    # via
    #   -r elastic_7x.in
    #   django-elasticsearch-dsl
factory_boy==2.11.1
    # via -r test.in
faker==8.1.4
    # via
    #   -r test.in
    #   factory-boy
filelock==3.0.12
    # via
    #   tox
    #   virtualenv
iniconfig==1.1.1
    # via pytest
mock==4.0.3
    # via -r test.in
packaging==21.0
    # via
    #   pytest
    #   tox
pillow==8.3.1
    # via -r common.in
platformdirs==2.1.0
    # via virtualenv
pluggy==0.13.1
    # via
    #   pytest
    #   tox
py==1.10.0
    # via
    #   -r test.in
    #   pytest
    #   tox
pyparsing==2.4.7
    # via packaging
pytest==6.2.4
    # via
    #   -r test.in
    #   pytest-cov
    #   pytest-django
    #   pytest-ordering
    #   pytest-pythonpath
pytest-cov==2.12.0
    # via -r test.in
pytest-django==4.3.0
    # via -r test.in
pytest-ordering==0.6
    # via -r test.in
pytest-pythonpath==0.7.3
    # via -r test.in
python-dateutil==2.8.2
    # via
    #   elasticsearch-dsl
    #   faker
python-memcached==1.58
    # via -r common.in
pytz==2021.1
    # via
    #   -r common.in
    #   django
selenium==3.141.0
    # via -r test.in
six==1.16.0
    # via
    #   -r common.in
    #   django-debug-toolbar-force
    #   django-elasticsearch-dsl
    #   django-ormex
    #   elasticsearch-dsl
    #   python-dateutil
    #   python-memcached
    #   tox
    #   virtualenv
sqlparse==0.4.1
    # via
    #   django
    #   django-debug-toolbar
text-unidecode==1.3
    # via faker
toml==0.10.2
    # via
    #   coverage
    #   pytest
    #   tox
tox==3.23.1
    # via -r test.in
urllib3==1.26.6
    # via
    #   elasticsearch
    #   selenium
virtualenv==20.6.0
    # via tox
